requirements_dev.txt
# package runtime requirements
python-dateutil==2.8.1
convertdate==2.2.0
korean_lunar_calendar==0.2.1
six==1.15.0

# test requirements
pytest
pytest-cov
coverage[toml]
pre-commit
flake8

#deployment
pip>=19.2.3
bump2version
wheel>=0.33.6
